Create Mission-oriented policies for lowcarbon futures in complex technological landscapes
This is the code for an agent-based model replicating the search and competition process of manufacturing firms. Technologies are embedded in an NK landscape (actually there are 2, one for costs and another for carbon intensity). They are engaged in price competition, but consumers have also fixed preferences for low-carbon products. The model replicates the main stylized facts observed in oligopolistic industries. It explores several policies, mainly R&D subsidies, adoption subsidies, carbon taxes, command and control, and research parks.